LS optimization, id fixes
