+++ apps/frontend/app/(website)/studio/main/ASTViewer/AboristViewer.tsx
@@ -1,109 +1,19 @@
 import { cn } from "@/utils";
-import type { Node } from "@babel/types";
 import Text from "@studio/components/Text";
 import { useScrollNodeIntoView } from "@studio/main/ASTViewer/useScrollNodeIntoView";
-import { extractComments } from "@studio/main/ASTViewer/utils";
 import {
-  useSelectFirstTreeNodeForSnippet,
+  EditorType,
   useSnippetsStore,
 } from "@studio/store/snippets";
 import { useRangesOnTarget } from "@studio/store/utils/useRangesOnTarget";
-import mapBabelASTToRenderableTree from "@studio/utils/mappers";
 import type React from "react";
 import { useCallback, useEffect, useRef, useState } from "react";
 import { type NodeApi, Tree } from "react-arborist";
 import useResizeObserver from "use-resize-observer";
+import { TreeNode } from "@studio/main/ASTViewer/utils";
+import { useSelectFirstTreeNodeForSnippet } from "@studio/main/ASTViewer/useSelectFirstTreeNodeForSnippet";
 
-type TreeNode = {
-  id: string;
-  actualNode: Node;
-  label: string;
-  children?: TreeNode[];
-  start: number;
-  end: number;
-};
 
-type EditorType = "before" | "after" | "output";
-
-interface Props {
-  type: EditorType;
-}
-
-const removeEmptyChildren = (node: TreeNode): TreeNode => {
-  const transformedChildren =
-    node.children
-      ?.map(removeEmptyChildren)
-      .filter((child) => child.children?.length || !child.children) || [];
-
-  return {
-    ...node,
-    children: transformedChildren.length > 0 ? transformedChildren : undefined,
-  };
-};
-
-const parseNodeData =
-  (node: TreeNode, identifier: string) => (param, index) => ({
-    id: `${node.id}-${identifier}-${index}`,
-    actualNode: param,
-    label: param.type,
-    start: param.start,
-    end: param.end,
-  });
-const getPosition = (node: TreeNode) => `${node.start}-${node.end}`;
-const transformTreeData = (node: TreeNode, relation?: string): TreeNode => {
-  const transformedNode: TreeNode = {
-    ...node,
-    relation,
-    children: [],
-  };
-
-  if (node.actualNode.type === "CallExpression") {
-    const args = node.actualNode.arguments.map(parseNodeData(node, "arg"));
-    const argumentsIds = args.map(getPosition);
-    return {
-      ...transformedNode,
-      children:
-        node.children.map((child) => {
-          const isParam = argumentsIds.includes(getPosition(child));
-          return transformTreeData(child, isParam ? "argument" : child.label);
-        }) ?? [],
-    };
-  }
-  if (node.actualNode.type === "FunctionDeclaration") {
-    const params = node.actualNode.params.map(parseNodeData(node, "param"));
-    const paramsIds = params.map(getPosition);
-    const children =
-      node.children.map((child) => {
-        const isParam = paramsIds.includes(getPosition(child));
-        const nameMap = {
-          BlockStatement: "body",
-          Identifier: "id",
-        };
-        return transformTreeData(
-          child,
-          (isParam ? "param" : nameMap[child.label]) ?? child.label,
-        );
-      }) ?? [];
-    return {
-      ...transformedNode,
-      children,
-    };
-  }
-
-  const children =
-    node.children?.map((child) =>
-      transformTreeData(
-        child,
-        Array.isArray(node.actualNode[node.label])
-          ? node.label
-          : Object.keys(node.actualNode).find(
-              (key) => node.actualNode[key] === child.actualNode,
-            ),
-      ),
-    ) ?? [];
-
-  return { ...transformedNode, children };
-};
 
 type Node = {
   id?: string;
@@ -115,18 +25,18 @@ type Node = {
   [key: string]: any;
 };
 
-export const ASTViewer: React.FC<Props> = ({ type }) => {
-  const ASTTreeRef = useRef<HTMLDivElement>(null);
+export const ASTViewer = ({ type }: {type: EditorType}) => {
+  const ASTTreeRef = useRef<HTMLDivElement | null>(null);
   const { width = 0, height = 0 } = useResizeObserver({ ref: ASTTreeRef });
   const getFirstTreeNode = useSelectFirstTreeNodeForSnippet();
   const [firstNode, setFirstNode] = useState<TreeNode | null>(null);
   const { getSelectedEditors } = useSnippetsStore();
   const {
-    [type]: { rootNode },
+    [type]: { rootNode, ranges },
   } = getSelectedEditors();
 
   const setRangesOnTarget = useRangesOnTarget();
-  const scrollNodeIntoView = useScrollNodeIntoView(type);
+  const scrollNodeIntoView = useScrollNodeIntoView();
 
   const handleNodeClick = useCallback(
     (node: NodeApi<TreeNode> = rootNode) => {
@@ -146,46 +56,54 @@ export const ASTViewer: React.FC<Props> = ({ type }) => {
     [rootNode, scrollNodeIntoView, setRangesOnTarget, type, getSelectedEditors],
   );
 
+  console.log({ranges})
   useEffect(() => {
+    console.log('useEffect')
     const firstTreeNode = getFirstTreeNode(type);
     if (firstTreeNode) {
       scrollNodeIntoView(firstTreeNode, ASTTreeRef);
       setFirstNode(firstTreeNode);
     }
-  }, [scrollNodeIntoView, getFirstTreeNode, type]);
+  }, [ranges, scrollNodeIntoView, getFirstTreeNode, type]);
 
   const NodeComponent = ({ node, style, dragHandle }) => {
     const isSelected = firstNode?.id === node.data.id;
     return (
-      <div style={style} ref={dragHandle}>
+      <div style={style} ref={dragHandle}
+           onClick={ (e) => {
+             e.stopPropagation();
+             handleNodeClick(node);
+           } }
+      >
+
         <Text
           className="cursor-pointer whitespace-nowrap"
-          color={isSelected ? "text-cyan-500" : undefined}
+          color={ isSelected ? "text-cyan-500" : undefined }
         >
-          {!node.isLeaf ? (
+          { !node.isLeaf ? (
             <strong
-              onClick={() => node.toggle()}
-              className={cn(node.isOpen ? "text-red-600" : "text-green-500")}
+              onClick={ (e) => {
+                node.toggle();
+                e.stopPropagation();
+              }
+            }
+              className={ cn(node.isOpen ? "text-red-600" : "text-green-500") }
             >
-              {node.isOpen ? "- " : "+ "}
+              { node.isOpen ? "- " : "+ " }
             </strong>
           ) : (
             <span>&nbsp;&nbsp;</span>
-          )}
+          ) }
           <span
-            onClick={(e) => {
-              e.stopPropagation();
-              handleNodeClick(node);
-            }}
             className="inline-block"
-            id={node.data.id}
+            id={ node.data.id }
           >
-            {node.data.relation && (
+            { node.data.relation && (
               <span className="text-purple-500 mr-1">
-                {node.data.relation}:{" "}
+                { node.data.relation }:{ " " }
               </span>
-            )}
-            {node.data.label}
+            ) }
+            { node.data.label }
           </span>
         </Text>
       </div>
@@ -202,10 +120,7 @@ export const ASTViewer: React.FC<Props> = ({ type }) => {
     );
   }
 
-  const transformedRootNode =
-    transformTreeData(mapBabelASTToRenderableTree(rootNode)) ?? null;
-  const withComments = removeEmptyChildren(transformedRootNode);
-  const { cleanedStructure } = extractComments(withComments);
+
 
   return (
     <div
@@ -213,7 +128,7 @@ export const ASTViewer: React.FC<Props> = ({ type }) => {
       ref={ASTTreeRef}
     >
       <Tree
-        data={cleanedStructure?.children ?? []}
+        data={rootNode?.children ?? []}
         openByDefault={true}
         width={width}
         height={height}
