+++ apps/frontend/app/(website)/studio/main/ASTViewer/useSelectFirstTreeNodeForSnippet.ts
@@ -1,0 +1,10 @@
+import { EditorType, useSnippetsStore } from "@studio/store/snippets";
+
+export const useSelectFirstTreeNodeForSnippet = () => {
+  const { getSelectedEditors } = useSnippetsStore();
+
+  return (type: EditorType) => {
+    const firstRange = getSelectedEditors()[type].ranges[0];
+    return firstRange && "id" in firstRange ? firstRange : null;
+  };
+};
