+++ apps/frontend/app/(website)/studio/src/store/mod.ts
@@ -8,7 +8,12 @@ import { persist } from "zustand/middleware";
 import { parseSnippet } from "../utils/babelParser";
 import mapBabelASTToRenderableTree from "../utils/mappers";
 import { type RangeCommand, buildRanges } from "../utils/tree";
-import { INITIAL_STATE } from "./initialState";
+import {
+  DEFAULT_FIND_REPLACE_EXPRESSION,
+  STARTER_SNIPPET,
+  TSMORPH_STARTER_SNIPPET
+} from "./initialState";
+import type { KnownEngines } from "@codemod-com/utilities";
 
 type ModStateValues = {
   name: string;
@@ -28,11 +33,24 @@ type ModStateSetters = {
   setCurrentCommand: (command: ModState["command"]) => void;
 };
 
+export const buildDefaultCodemodSource = (engine: KnownEngines) => {
+  if (engine === "jscodeshift") {
+    return prettify(
+      STARTER_SNIPPET.replace(
+        "{%DEFAULT_FIND_REPLACE_EXPRESSION%}",
+        DEFAULT_FIND_REPLACE_EXPRESSION,
+      ).replace("{%COMMENT%}", ""),
+    );
+  }
+
+  return TSMORPH_STARTER_SNIPPET;
+};
+
 export type ModState = ModStateSetters & ModStateValues;
 const getInitialState = (): ModStateValues => {
   const savedState = isServer ? null : localStorage.getItem("mod-store");
-  const hasSavedState = savedState && JSON.parse(savedState).content;
-  const parsed = parseSnippet(hasSavedState || INITIAL_STATE.codemodSource);
+  const initialState = savedState ? JSON.parse(savedState).content : buildDefaultCodemodSource("jscodeshift");
+  const parsed = parseSnippet(initialState);
 
   const parsedContent = isFile(parsed)
     ? mapBabelASTToRenderableTree(parsed)
@@ -40,12 +58,12 @@ const getInitialState = (): ModStateValues => {
 
   return {
     name: "",
-    content: INITIAL_STATE.codemodSource,
+    content: initialState,
     hasRuntimeErrors: false,
     parsedContent,
     ranges: [],
     rangesUpdatedAt: Date.now(),
-    command: INITIAL_STATE.command,
+    command: null,
   };
 };
