+++ apps/frontend/app/(website)/studio/src/store/snippets.ts
@@ -1,16 +1,12 @@
-import { isFile } from "@babel/types";
 import type { KnownEngines } from "@codemod-com/utilities";
 import { isServer } from "@studio/config";
 import type { OffsetRange } from "@studio/schemata/offsetRangeSchemata";
-import { INITIAL_STATE } from "@studio/store/initialState";
-import { getSnippetInitialState } from "@studio/store/utils/getSnippetInitialState";
+import { getSingleTestCase, getSnippetInitialState } from "@studio/store/utils/getSnippetInitialState";
 import type { TreeNode } from "@studio/types/tree";
-import { parseSnippet } from "@studio/utils/babelParser";
-import mapBabelASTToRenderableTree from "@studio/utils/mappers";
-import { type RangeCommand, buildRanges } from "@studio/utils/tree";
-import { map, mapObjIndexed, reduce, remove } from "ramda";
+import { buildRanges, type RangeCommand } from "@studio/utils/tree";
+import { map, mapObjIndexed, omit, reduce, remove } from "ramda";
 import { create } from "zustand";
-import { persist } from "zustand/middleware";
+import { transformNode } from "@studio/main/ASTViewer/utils";
 
 export type Token = Readonly<{
   start: number;
@@ -31,20 +27,20 @@ type SnippetSetters = {
 };
 
 type SnippetValuesMap = {
-  [K in EditorType as `${K}Snippet`]: SnippetValues["content"];
+  [K in EditorType as `${ K }Snippet`]: SnippetValues["content"];
 };
 
 type SnippetSettersMap = {
-  [K in EditorType as `set${Capitalize<K>}Snippet`]: SnippetSetters["setContent"];
+  [K in EditorType as `set${ Capitalize<K> }Snippet`]: SnippetSetters["setContent"];
 } & {
-  [K in EditorType as `set${Capitalize<K>}Selection`]: SnippetSetters["setSelection"];
+  [K in EditorType as `set${ Capitalize<K> }Selection`]: SnippetSetters["setSelection"];
 };
 type SnippetsConfig = {
   getHasReachedTabsLimit: () => boolean;
   tabsLimit: number;
   currentContent: string;
   currentType: EditorType;
-  addPair: () => void;
+  addPair: (name?: string) => void;
   clearAll: () => void;
   setInitialState: (state: Partial<SnippetsState>) => void;
   removePair: (index: number) => void;
@@ -53,11 +49,11 @@ type SnippetsConfig = {
   getSelectedEditors: () => Editors &
     SnippetValuesMap &
     SnippetSettersMap & {
-      setSelection: (x: EditorType) => (command: RangeCommand) => void;
-      setRanges: (x: EditorType) => (command: RangeCommand) => void;
-      setContent: (x: EditorType) => (command: RangeCommand) => void;
-      renameEditor: (index: number) => (name: string) => void;
-    };
+    setSelection: (x: EditorType) => (command: RangeCommand) => void;
+    setRanges: (x: EditorType) => (command: RangeCommand) => void;
+    setContent: (x: EditorType) => (command: RangeCommand) => void;
+    renameEditor: (index: number) => (name: string) => void;
+  };
   setEngine: (engine: KnownEngines) => void;
   setSelectedPairIndex: (index: number) => void;
 };
@@ -82,8 +78,8 @@ const toEditorSnippets = (editors: Editors): EditorsSnippets => ({
 type AllEditors = {
   [x in keyof Editors]: SnippetValues[];
 };
-type EditorType = keyof Editors;
-type AllSnippets = {
+export type EditorType = keyof Editors;
+export type AllSnippets = {
   before: string[];
   after: string[];
   output: string[];
@@ -109,173 +105,175 @@ const getEditorsFromLS = () => {
   return;
 };
 export const useSnippetsStore = create<SnippetsState>(
-  persist(
-    (set, get) => ({
-      tabsLimit: 12,
-      getHasReachedTabsLimit: () => get().editors.length >= get().tabsLimit,
-      editors: INITIAL_STATE.editors,
-      addPair: () => {
-        set({
-          editors: [
-            ...get().editors,
-            {
-              name: `Test ${
-                (get()
-                  .getAllNames()
-                  .filter((name) => name.toLowerCase().startsWith("test "))
-                  .map((name) => name.split(" ")[1])
-                  .map(Number)
-                  .filter(Boolean)
-                  .at(-1) || 0) + 1
-              }`,
-              before: getSnippetInitialState(),
-              after: getSnippetInitialState(),
-              output: getSnippetInitialState(),
-            },
-          ],
-        });
-        setTimeout(
-          () =>
+    (set, get) =>
+      ({
+        tabsLimit: 12,
+        getHasReachedTabsLimit: () => get().editors.length >= get().tabsLimit,
+        // editors: localStorage.getItem('editors') ? JSON.parse(localStorage.getItem('editors')) : [getSingleTestCase()],
+        addPair: (name?: string) => {
+          set({
+            editors: [
+              ...get().editors,
+              {
+                name: name ||`Test ${
+                  (get()
+                    .getAllNames()
+                    .filter((name) => name.toLowerCase().startsWith("test "))
+                    .map((name) => name.split(" ")[1])
+                    .map(Number)
+                    .filter(Boolean)
+                    .at(-1) || 0) + 1
+                }`,
+                before: getSnippetInitialState(),
+                after: getSnippetInitialState(),
+                output: getSnippetInitialState(),
+              },
+            ],
+          });
+          setTimeout(
+            () =>
+              set({
+                selectedPairIndex: get().editors.length - 1,
+              }),
+            100,
+          );
+        },
+        editors: [],
+        renameEditor: (index) => (name) => {
+          const obj = get();
+          obj.editors[index].name = name;
+          set(obj);
+        },
+        removePair: (index: number) => {
+          const editors =
+            get().editors.length > 1
+              ? remove(index, 1, get().editors)
+              : get().editors;
+          if (index === get().selectedPairIndex) {
             set({
-              selectedPairIndex: get().editors.length - 1,
-            }),
-          100,
-        );
-      },
-      renameEditor: (index) => (name) => {
-        const obj = get();
-        obj.editors[index].name = name;
-        set(obj);
-      },
-      removePair: (index: number) => {
-        const editors =
-          get().editors.length > 1
-            ? remove(index, 1, get().editors)
-            : get().editors;
-        if (index === get().selectedPairIndex) {
+              selectedPairIndex: 0,
+              editors,
+            });
+          } else
+            set({
+              selectedPairIndex: Math.max(get().selectedPairIndex - 1, 0),
+              editors,
+            });
+        },
+        clearAll: () => {
           set({
             selectedPairIndex: 0,
-            editors,
           });
-        } else
+          setTimeout(
+            () =>
+              set({
+                editors: [
+                  {
+                    name: "Test 1",
+                    before: getSnippetInitialState(),
+                    after: getSnippetInitialState(),
+                    output: getSnippetInitialState(),
+                  },
+                ],
+              }),
+            100,
+          );
+        },
+        selectedPairIndex: 0,
+        getAllNames: () => get().editors.map(({ name }) => name),
+        getAllSnippets: () => {
+          return mapObjIndexed(
+            map(({ content }: SnippetValues) => content),
+            reduce(
+              (acc, { before, after, output }) => ({
+                before: [...acc.before, before],
+                after: [...acc.after, after],
+                output: [...acc.output, output],
+              }),
+              {
+                before: [],
+                after: [],
+                output: [],
+              } as AllEditors,
+              get().editors,
+            ),
+          );
+        },
+        setSelectedPairIndex: (i: number) => {
+          set({ selectedPairIndex: i });
+        },
+        getSelectedEditors: () => {
+          const index = get().selectedPairIndex || 0;
+          const editors = get().editors?.[index] as Editors;
+          return {
+            ...editors,
+            setContent: (type) => get().setContent(index, type),
+            beforeSnippet: editors?.before?.content || "",
+            afterSnippet: editors?.after?.content || "",
+            outputSnippet: editors?.output?.content || "",
+            setBeforeSnippet: get().setContent(index, "before"),
+            setAfterSnippet: get().setContent(index, "after"),
+            setOutputSnippet: get().setContent(index, "output"),
+            setBeforeSelection: get().setSelection(index, "before"),
+            setAfterSelection: get().setSelection(index, "after"),
+            setOutputSelection: get().setSelection(index, "output"),
+            setSelection: (editorType: EditorType) =>
+              get().setSelection(index, editorType),
+          };
+        },
+        setInitialState: (state) => {
+          set(state);
+        },
+        setEngine: (engine) => {
+          localStorage.setItem(
+            "engine",
+            engine,
+          );
           set({
-            selectedPairIndex: Math.max(get().selectedPairIndex - 1, 0),
-            editors,
-          });
-      },
-      clearAll: () => {
-        set({
-          selectedPairIndex: 0,
-        });
-        setTimeout(
-          () =>
-            set({
-              editors: [
-                {
-                  name: "Test 1",
-                  before: getSnippetInitialState(),
-                  after: getSnippetInitialState(),
-                  output: getSnippetInitialState(),
-                },
-              ],
-            }),
-          100,
-        );
-      },
-      engine: INITIAL_STATE.engine,
-      selectedPairIndex: 0,
-      getAllNames: () => get().editors.map(({ name }) => name),
-      getAllSnippets: () => {
-        return mapObjIndexed(
-          map(({ content }: SnippetValues) => content),
-          reduce(
-            (acc, { before, after, output }) => ({
-              before: [...acc.before, before],
-              after: [...acc.after, after],
-              output: [...acc.output, output],
-            }),
-            {
-              before: [],
-              after: [],
-              output: [],
-            } as AllEditors,
-            get().editors,
-          ),
-        );
-      },
-      setSelectedPairIndex: (i: number) => {
-        set({ selectedPairIndex: i });
-      },
-      getSelectedEditors: () => {
-        const index = get().selectedPairIndex || 0;
-        const editors = get().editors?.[index] as Editors;
-        return {
-          ...editors,
-          setContent: (type) => get().setContent(index, type),
-          beforeSnippet: editors?.before?.content || "",
-          afterSnippet: editors?.after?.content || "",
-          outputSnippet: editors?.output?.content || "",
-          setBeforeSnippet: get().setContent(index, "before"),
-          setAfterSnippet: get().setContent(index, "after"),
-          setOutputSnippet: get().setContent(index, "output"),
-          setBeforeSelection: get().setSelection(index, "before"),
-          setAfterSelection: get().setSelection(index, "after"),
-          setOutputSelection: get().setSelection(index, "output"),
-          setSelection: (editorType: EditorType) =>
-            get().setSelection(index, editorType),
-        };
-      },
-      setInitialState: (state) => {
-        set(state);
-      },
-      setEngine: (engine) =>
-        set({
-          engine,
-        }),
-      setContent: (editorsPairIndex, type) => {
-        return (content) => {
-          const parsed = parseSnippet(content);
-          const rootNode = isFile(parsed)
-            ? mapBabelASTToRenderableTree(parsed)
-            : null;
+            engine,
+          })
+        },
+        setContent: (editorsPairIndex, type) => {
+          return (content) => {
 
-          const obj = get();
-          obj.editors[editorsPairIndex][type].content = content;
-          obj.editors[editorsPairIndex][type].rootNode = rootNode;
-          set({ currentContent: content, currentType: type, ...obj });
-          try {
-            localStorage.setItem(
-              "editors",
-              JSON.stringify(obj.editors.map(toEditorSnippets)),
-            );
-          } catch (error) {
-            console.error("error on JSON.stringify(obj.editors) ", { error });
+            const obj = get();
+            obj.editors[editorsPairIndex][type].content = content;
+            obj.editors[editorsPairIndex][type].rootNode = transformNode(content, type);
+            set({ currentContent: content, currentType: type, ...obj });
+            try {
+              localStorage.setItem(
+                "editors",
+                JSON.stringify(obj.editors.map(toEditorSnippets)),
+              );
+            } catch (error) {
+              console.error("error on JSON.stringify(obj.editors) ", { error });
+            }
+          };
+        },
+        setEditors: (editorsContents: EditorsSnippets[]) => {
+          editorsContents.forEach((eC, i) => {
+              get().addPair(eC.name);
+              Object.entries(omit(['name'], eC))
+                .forEach(([propName, editorContent]) => {
+                  get().setContent(i, propName)(editorContent);
+                  get().setSelection(i, propName)({ kind: "PASS_THROUGH", ranges: [] });
+                })
+            }
+          )
+        },
+        setSelection: (editorsPairIndex, type) => (command) => {
+          const rootNode = get().editors[editorsPairIndex]?.[type]?.rootNode;
+          if (rootNode) {
+            const ranges = buildRanges(rootNode, command);
+
+            const obj = get();
+            obj.editors[editorsPairIndex][type].ranges = ranges;
+            obj.editors[editorsPairIndex][type].rangeUpdatedAt = Date.now();
+            set(obj);
           }
-        };
-      },
-      setSelection: (editorsPairIndex, type) => (command) => {
-        const rootNode = get().editors[editorsPairIndex]?.[type]?.rootNode;
-        if (rootNode) {
-          const ranges = buildRanges(rootNode, command);
+        },
+      }),
+  )
 
-          const obj = get();
-          obj.editors[editorsPairIndex][type].ranges = ranges;
-          obj.editors[editorsPairIndex][type].rangeUpdatedAt = Date.now();
-          set(obj);
-        }
-      },
-    }),
-    {
-      name: "snippets-storage",
-    },
-  ),
-);
-
-export const useSelectFirstTreeNodeForSnippet = () => {
-  const { getSelectedEditors } = useSnippetsStore();
-
-  return (type: EditorType) => {
-    const firstRange = getSelectedEditors()[type].ranges[0];
-    return firstRange && "id" in firstRange ? firstRange : null;
-  };
-};
+useSnippetsStore?.getState?.().setEditors(localStorage.getItem('editors') ? JSON.parse(localStorage.getItem('editors')) : [getSingleTestCase()])
+useSnippetsStore?.getState?.().setEngine(localStorage.getItem('engine') as KnownEngines || "jscodeshift")
