+++ apps/frontend/app/(website)/studio/src/store/utils/getSnippetInitialState.ts
@@ -6,6 +6,7 @@ import {
 import type { SnippetValues } from "@studio/store/snippets";
 import { parseSnippet } from "@studio/utils/babelParser";
 import mapBabelASTToRenderableTree from "@studio/utils/mappers";
+import { transformNode } from "@studio/main/ASTViewer/utils";
 
 export const toInitialStates = ({ before, after, name }) => ({
   name,
@@ -19,9 +20,7 @@ export const getSnippetInitialState = (
 ): SnippetValues => {
   const content = defaultContent ?? "";
   const contentParsed = parseSnippet(content);
-  const rootNode = isFile(contentParsed)
-    ? mapBabelASTToRenderableTree(contentParsed)
-    : null;
+  const rootNode = transformNode(contentParsed)
 
   const tokens: SnippetValues["tokens"] = isFile(contentParsed)
     ? Array.isArray(contentParsed.tokens)
